mover AboutScreen para components/screens
