+++ pages/sobre.js
@@ -1,11 +1,7 @@
-import React from 'react';
-
-import PropTypes from 'prop-types';
 import { GraphQLClient, gql } from 'graphql-request';
-import Box from '../src/components/foundation/layout/Box';
-import Grid from '../src/components/foundation/layout/Grid';
-import Text from '../src/components/foundation/Text';
+
 import websitePageHOC from '../src/components/wrappers/WebsitePage/hoc';
+import AboutScreen from '../src/components/screens/AboutScreen';
 
 export async function getStaticProps() {
   const TOKEN = process.env.DATO_READONLY_TOKEN;
@@ -34,48 +30,5 @@ export async function getStaticProps() {
     },
   };
 }
-
-function AboutScreen({ messages }) {
-  return (
-    <Box
-      display="flex"
-      flexDirection="column"
-      flex={1}
-    >
-      <Grid.Container>
-        <Grid.Row
-          marginTop={{ xs: '32px', md: '120px' }}
-          flex="1"
-        >
-          <Grid.Col
-            value={{ xs: 12, md: 6, lg: 6 }}
-            offset={{ md: 2 }}
-            flex={1}
-          >
-            <Text
-              variant="title"
-              tag="h2"
-              color="tertiary.main"
-            >
-              {messages.pageSobre.pageTitle}
-            </Text>
-
-            <Box
-              dangerouslySetInnerHTML={{
-                __html: messages.pageSobre.pageDescription,
-              }}
-            />
-
-          </Grid.Col>
-        </Grid.Row>
-      </Grid.Container>
-    </Box>
-  );
-}
-
-AboutScreen.propTypes = {
-  // eslint-disable-next-line react/forbid-prop-types
-  messages: PropTypes.object.isRequired,
-};
 
 export default websitePageHOC(AboutScreen);
